Split init game map fleet markers into part file
Moves fleet tile marker helpers to init_game_map_view_builder_fleet_markers.dart
so init_game_map_view_builder.dart stays under the 1000 non-comment line
repo-lint limit (repo.dart_file_non_comment_line_size).
